+++ registryWorkers.go
@@ -1,16 +1,16 @@
 package kratos
 
 import (
+	"context"
 	"fmt"
 	"net/http"
 	"sync"
 	"sync/atomic"
 
-	"github.com/go-kit/kit/log"
 	"github.com/goph/emperror"
-	"github.com/xmidt-org/webpa-common/logging"
-	"github.com/xmidt-org/webpa-common/semaphore"
 	"github.com/xmidt-org/wrp-go/v3"
+	"go.uber.org/zap"
+	"golang.org/x/sync/semaphore"
 )
 
 // registryHandler is a way to send the wrp message to the correct handler.
@@ -29,16 +29,16 @@ type registryQueue struct {
 	sendFunc         sendWRPFunc
 	downstreamSender downstreamSender
 	deviceID         string
-	workers          semaphore.Interface
+	workers          *semaphore.Weighted
 	wg               sync.WaitGroup
-	logger           log.Logger
+	logger           *zap.Logger
 	once             sync.Once
 	closed           atomic.Value
 }
 
 // NewRegistryHandler returns a registryHandler, which sends wrp messages to
 // the correct handler in an asynchronous fashion.
-func NewRegistryHandler(senderFunc sendWRPFunc, registry HandlerRegistry, downstreamSender downstreamSender, maxWorkers int, queueSize int, deviceID string, logger log.Logger) *registryQueue {
+func NewRegistryHandler(senderFunc sendWRPFunc, registry HandlerRegistry, downstreamSender downstreamSender, maxWorkers int, queueSize int, deviceID string, logger *zap.Logger) *registryQueue {
 	size := queueSize
 	if size < minQueueSize {
 		size = minQueueSize
@@ -53,7 +53,7 @@ func NewRegistryHandler(senderFunc sendWRPFunc, registry HandlerRegistry, downst
 		sendFunc:         senderFunc,
 		downstreamSender: downstreamSender,
 		deviceID:         deviceID,
-		workers:          semaphore.New(numWorkers),
+		workers:          semaphore.NewWeighted(int64(numWorkers)),
 		logger:           logger,
 	}
 	r.wg.Add(1)
@@ -88,9 +88,10 @@ func (r *registryQueue) Close() {
 // registryQueue to read from its queue, get the appropriate handler for the
 // given message, and send it using the downstreamSender.
 func (r *registryQueue) startGettingHandlers() {
+	ctx := context.Background()
 	defer r.wg.Done()
 	for i := range r.incoming {
-		r.workers.Acquire()
+		r.workers.Acquire(ctx, 1)
 		r.wg.Add(1)
 		go r.getHandler(i)
 	}
@@ -100,29 +101,27 @@ func (r *registryQueue) startGettingHandlers() {
 // the message.
 func (r *registryQueue) getHandler(msg *wrp.Message) {
 	defer r.wg.Done()
-	defer r.workers.Release()
+	defer r.workers.Release(1)
 
-	logging.Debug(r.logger).Log(logging.MessageKey(), "Getting handler...")
+	r.logger.Debug("Getting handler...")
 
 	handler, err := r.registry.GetHandler(msg.Destination)
 	if _, ok := err.(ErrNoDownstreamHandler); ok {
 		// If no valid handlers for the destination, create a new simple RequestResponse wrp with http Status Code of Service Unavailable
 		response := CreateErrorWRP(msg.TransactionUUID, msg.Source, r.deviceID, http.StatusServiceUnavailable, emperror.Wrap(err, "unable to get handler"))
-		logging.Error(r.logger, emperror.Context(err)...).
-			Log(logging.MessageKey(), "Failed to get handler", logging.ErrorKey(), err.Error())
+		r.logger.Error("Failed to get handler", zap.Error(err))
 		r.sendFunc(response)
 		return
 	}
 	if err != nil {
 		// for now, do the same as if there is no downstream handler.
 		response := CreateErrorWRP(msg.TransactionUUID, msg.Source, r.deviceID, http.StatusServiceUnavailable, emperror.Wrap(err, "unable to get handler"))
-		logging.Error(r.logger, emperror.Context(err)...).
-			Log(logging.MessageKey(), "Failed to get handler", logging.ErrorKey(), err.Error())
+		r.logger.Error("Failed to get handler", zap.Error(err))
 		r.sendFunc(response)
 		return
 	}
 
 	r.downstreamSender.Send(handler, msg)
 
-	logging.Debug(r.logger).Log(logging.MessageKey(), "Sent message to handler")
+	r.logger.Debug("Sent message to handler")
 }